backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
motor==3.7.1
python-dotenv==1.1.0
pydantic==2.11.5
python-multipart==0.0.6
email-validator==2.1.0.post1 